+++ crates/common/src/provider/mod.rs
@@ -2,5 +2,5 @@
 
 pub mod alloy;
 pub mod ethers;
-pub mod policy;
+pub mod retry;
 pub mod runtime_transport;
